@@ -1,30 +1,32 @@
 name: flutter_trip
 description: A new Flutter application.
 
-version: 1.0.0+1
+version: 1.0.1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^0.1.3
   # banner
-  flutter_swiper: ^1.1.4
+  flutter_swiper: ^1.1.6
   # http
-  http: ^0.12.0+1
+  http: ^0.12.2
   # webview
-  flutter_webview_plugin: ^0.3.0+2
+  webview_flutter: ^2.0.8
+#  flutter_webview_plugin: ^0.4.0
   # 瀑布流
-  flutter_staggered_grid_view: ^0.3.0
+  flutter_staggered_grid_view: ^0.3.4
   # 启动白屏
   #  flutter_splash_screen: ^0.3.0
   #  flare_splash_screen: ^4.0.0
-
   # taost
-  fluttertoast: ^8.0.7
+#  fluttertoast: ^8.0.9
+  # placeholder
+  transparent_image: ^2.0.0
 
 dev_dependencies:
   flutter_test:
