@@ -1,12 +1,12 @@
 class CommonModel {
-  final String icon;
-  final String title;
+  final String? icon;
+  final String? title;
   final String url;
-  final String statusBarColor;
-  final bool hideAppBar;
+  final String? statusBarColor;
+  final bool? hideAppBar;
 
   CommonModel(
-      {this.icon, this.title, this.url, this.statusBarColor, this.hideAppBar});
+      {this.icon, this.title, required this.url, this.statusBarColor, this.hideAppBar});
 
   factory CommonModel.fromJson(Map<String, dynamic> json) {
     return CommonModel(
