@@ -4,17 +4,17 @@
 
 ///旅拍类别模型
 class TravelTabModel {
-  Map params;
-  String url;
-  List<TravelTab> tabs;
+  late Map params;
+  late String url;
+  late List<TravelTab> tabs;
 
-  TravelTabModel({this.url, this.tabs});
+  TravelTabModel(this.url, this.tabs);
 
   TravelTabModel.fromMap(Map<String, dynamic> json) {
     url = json['url'];
     params = json['params'];
     if (json['tabs'] != null) {
-      tabs = new List<TravelTab>();
+      tabs = new List<TravelTab>.empty(growable: true);
       json['tabs'].forEach((v) {
         tabs.add(new TravelTab.fromJson(v));
       });
@@ -32,10 +32,10 @@ class TravelTabModel {
 }
 
 class TravelTab {
-  String labelName;
-  String groupChannelCode;
+  late String labelName;
+  late String groupChannelCode;
 
-  TravelTab({this.labelName, this.groupChannelCode});
+  TravelTab(this.labelName, this.groupChannelCode);
 
   TravelTab.fromJson(Map<String, dynamic> json) {
     labelName = json['labelName'];
