@@ -1,13 +1,13 @@
 class ConfigModel {
   final String searchUrl;
 
-  ConfigModel({this.searchUrl});
+  ConfigModel({required this.searchUrl});
 
   factory ConfigModel.fromJson(Map<String, dynamic> json) {
     return ConfigModel(searchUrl: json['searchUrl']);
   }
 
   Map<String, dynamic> toJson() {
-    return {searchUrl: searchUrl};
+    return {searchUrl!: searchUrl};
   }
 }
