@@ -6,15 +6,14 @@ class GridNavModel {
   final GridNavItem flight;
   final GridNavItem travel;
 
-  GridNavModel({this.hotel, this.flight, this.travel});
+  GridNavModel(
+      {required this.hotel, required this.flight, required this.travel});
 
   factory GridNavModel.fromJson(Map<String, dynamic> json) {
-    return json != null
-        ? GridNavModel(
-            hotel: GridNavItem.fromJson(json['hotel']),
-            flight: GridNavItem.fromJson(json['flight']),
-            travel: GridNavItem.fromJson(json['travel']))
-        : null;
+    return GridNavModel(
+        hotel: GridNavItem.fromJson(json['hotel']),
+        flight: GridNavItem.fromJson(json['flight']),
+        travel: GridNavItem.fromJson(json['travel']));
   }
 }
 
@@ -28,13 +27,13 @@ class GridNavItem {
   final CommonModel item4;
 
   GridNavItem(
-      {this.startColor,
-      this.endColor,
-      this.mainItem,
-      this.item1,
-      this.item2,
-      this.item3,
-      this.item4});
+      {required this.startColor,
+      required this.endColor,
+      required this.mainItem,
+      required this.item1,
+      required this.item2,
+      required this.item3,
+      required this.item4});
 
   factory GridNavItem.fromJson(Map<String, dynamic> json) {
     return GridNavItem(
