@@ -2,18 +2,19 @@
 /// totalCount : 500
 /// resultList : [{"type":1,"article":{"articleId":4753554,"articleType":"star","productType":5,"sourceType":126,"articleTitle":"【等到樱花烂漫盛开时】","author":{"authorId":430224,"nickName":"程晓玥YvonneChing","clientAuth":"E5BFB9A6F8D62244C5644C8B54A9271C","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=E5BFB9A6F8D62244C5644C8B54A9271C","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80v10000000oofb338E2_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80v10000000oofb338E2_C_180_180.jpg"},"identityType":1,"tag":""},"images":[{"imageId":309579089,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0n12000000sei6395D2_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0n12000000sei6395D2.jpg","width":1280,"height":1920,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":4970,"likeCount":3071,"commentCount":1666,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4753554.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4753554.html","wxUrl":"/pages/article/detail?articleId=4753554"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc1MzU1NA==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4753554","wxUrl":"/pages/detail/detail?articleId=4753554"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc1MzU1NA==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4753554","wxUrl":"/pages/you/lvpai/detail?articleId=4753554"}],"tags":[],"topics":[{"topicId":6544,"topicName":"踏青赏花正当时","level":1},{"topicId":3100,"topicName":"春天去哪玩","level":1},{"topicId":12,"topicName":"旅行酷体验","level":1},{"topicId":30,"topicName":"网红打卡地","level":1},{"topicId":1005,"topicName":"旅拍美少女","level":1}],"pois":[{"poiType":3,"poiId":78941,"poiName":"上野公园","businessId":13354,"districtId":294,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=78941&poiType=3&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=78941&poiType=3&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":false,"countryName":"日本"},{"poiType":4,"poiId":294,"poiName":"东京","districtId":294,"districtName":"东京","districtENName":"Tokyo","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=294&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=294&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":false,"countryName":"日本"}],"publishTime":"2019-03-19 11:11:11","publishTimeDisplay":"2天前","shootTime":"2019-03-19 11:11:11","shootTimeDisplay":"2019-03-19","level":1,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":55,"articleStatus":2,"poiName":"上野公园"}},{"type":1,"article":{"articleId":4568357,"articleType":"movie","productType":5,"sourceType":126,"articleTitle":"拍照攻略大赛|去超市拍一组可爱的照片吧～ ‍‍‍ ………… ‍tips： 和各种吃的玩的互动","author":{"authorId":404904,"nickName":"Silvie旅行手帐","clientAuth":"05497A8B0B0D5B295C00D57B6AB6546E05FB737263792735DE8DAA87A76C93DF","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=05497A8B0B0D5B295C00D57B6AB6546E05FB737263792735DE8DAA87A76C93DF","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z8060z000000nd39sD266_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z8060z000000nd39sD266_C_180_180.jpg"},"tag":""},"images":[{"imageId":307271009,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0211000000rg4xlBAF9_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0211000000rg4xlBAF9.jpg","width":1280,"height":960,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":18612,"likeCount":3135,"commentCount":1486,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4568357.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4568357.html","wxUrl":"/pages/article/detail?articleId=4568357"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDU2ODM1Nw==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4568357","wxUrl":"/pages/detail/detail?articleId=4568357"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDU2ODM1Nw==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4568357","wxUrl":"/pages/you/lvpai/detail?articleId=4568357"}],"tags":[],"topics":[{"topicId":4132,"topicName":"拍照攻略大赛","level":1,"iconImage":{"imageId":306610393,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6.png","width":39,"height":45,"mediaType":0}},{"topicId":3096,"topicName":"一学就会拍照技术流","level":1},{"topicId":3100,"topicName":"春天去哪玩","level":1},{"topicId":1005,"topicName":"旅拍美少女","level":1},{"topicId":104,"topicName":"春节","level":2}],"pois":[{"poiType":4,"poiId":100009,"poiName":"四川","districtId":100009,"districtName":"四川","districtENName":"Sichuan","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=100009&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=100009&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"}],"publishTime":"2019-02-22 17:09:22","publishTimeDisplay":"27天前","shootTime":"2019-02-02 12:01:01","shootTimeDisplay":"2019-02-02","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":95,"articleStatus":2,"poiName":""}},{"type":1,"article":{"articleId":4198911,"articleType":"test","productType":5,"sourceType":126,"articleTitle":"国内首创1200平米，无边际泳池，70米高悬挑透明泳池，","author":{"authorId":405252,"nickName":"亼乐阳光","clientAuth":"E6620E1ABF4E93F96646A36DDAA60905","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=E6620E1ABF4E93F96646A36DDAA60905","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80o11000000qzwgh07B7_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80o11000000qzwgh07B7_C_180_180.jpg"},"tag":""},"images":[{"imageId":302091705,"dynamicUrl":"https://youimg1.c-ctrip.com/target/10010z000000nqn2m26D1_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/10010z000000nqn2m26D1.jpg","width":1203,"height":740,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":1401,"likeCount":16,"commentCount":0,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4198911.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4198911.html","wxUrl":"/pages/article/detail?articleId=4198911"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDE5ODkxMQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4198911","wxUrl":"/pages/detail/detail?articleId=4198911"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDE5ODkxMQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4198911","wxUrl":"/pages/you/lvpai/detail?articleId=4198911"}],"tags":[],"topics":[{"topicId":528,"topicName":"12月去哪儿","level":1},{"topicId":530,"topicName":"跨年好去处","level":1},{"topicId":522,"topicName":"景点打卡","level":1,"iconImage":{"imageId":302980116,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100b0y000000m6mqa07F4_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100b0y000000m6mqa07F4.png","width":48,"height":48,"mediaType":0}},{"topicId":30,"topicName":"网红打卡地","level":1}],"pois":[{"poiType":50,"poiId":35225106,"poiName":"三亚康年酒店","businessId":9789923,"districtId":61,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=35225106&poiType=50&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=35225106&poiType=50&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"},{"poiType":4,"poiId":61,"poiName":"三亚","districtId":61,"districtName":"三亚","districtENName":"Sanya","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=61&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=61&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":true,"countryName":"中国"}],"publishTime":"2018-12-20 01:25:16","publishTimeDisplay":"2018-12-20","shootTime":"2018-12-20 01:25:15","shootTimeDisplay":"2018-12-20","level":2,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":0,"articleStatus":2,"poiName":"三亚康年酒店"}},{"type":3,"advert":{"id":15872,"moduleId":776002,"appUrl":"/rn_destination_video/main.js?CRNModuleName=destinationlive&CRNType=1&initialPage=travelRecord_detail&id=4764456","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4764456&isHideHeader=true&isHideNavBar=YES&popup=close&autoawaken=close","wxUrl":"","imageUrl":"https://dimg02.c-ctrip.com/images/100312000000sk40pEC7D.jpg","width":512,"height":750}},{"type":1,"article":{"articleId":115519,"articleType":"hot","productType":5,"sourceType":5,"articleTitle":"罗平油菜花海#向往的生活","articleDesc":" 罗平是我国主要的油菜籽生产基地之一，和我去年...","author":{"authorId":242547,"nickName":"笑飞雪","clientAuth":"BDDB63B0F6C7CEDDEE2E3B0FF82D51CC","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=BDDB63B0F6C7CEDDEE2E3B0FF82D51CC","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80g0m000000dxm4079E2_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80g0m000000dxm4079E2_C_180_180.jpg"},"qualification":"https://pages.c-ctrip.com/you/note/balloon_author.png","identityType":4,"tag":""},"images":[{"imageId":104096046,"dynamicUrl":"https://youimg1.c-ctrip.com/target/10030b0000005rr1eF6C5_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/10030b0000005rr1eF6C5.jpg","width":1920,"height":1276,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":18838,"likeCount":3369,"commentCount":1859,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/115519.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/115519.html","wxUrl":"/pages/article/detail?articleId=115519"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9MTE1NTE5","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=115519","wxUrl":"/pages/detail/detail?articleId=115519"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9MTE1NTE5","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=115519","wxUrl":"/pages/you/lvpai/detail?articleId=115519"}],"tags":[{"tagId":197,"tagName":"赏油菜花","tagLevel":2,"parentTagId":88,"source":0,"sortIndex":12},{"tagId":222,"tagName":"植物","tagLevel":2,"parentTagId":122,"source":0,"sortIndex":0}],"topics":[{"topicId":467,"topicName":"花海","level":3},{"topicId":209,"topicName":"春日寻芳","level":2},{"topicId":215,"topicName":"赏花","level":2},{"topicId":726,"topicName":"在旅途拍世界","level":1},{"topicId":37,"topicName":"在当地过节","level":1}],"pois":[{"poiType":4,"poiId":676,"poiName":"罗平","districtId":676,"districtName":"罗平","districtENName":"Luoping","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=676&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=676&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"}],"publishTime":"2018-12-25 10:47:27","publishTimeDisplay":"2018-12-25","shootTime":"2018-12-25 10:47:27","shootTimeDisplay":"2018-12-25","sourceInfo":"9525","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":102,"articleStatus":2,"poiName":""}},{"type":1,"article":{"articleId":4765591,"articleType":"hot","productType":5,"sourceType":126,"articleTitle":"春暖花开时，我只想与你去看最美梯田花海 时隔三月，二入篁岭。若是冬日第一次的邂逅是浓烈热闹的晒秋，","author":{"authorId":20591,"nickName":"鱼鱼众生","clientAuth":"593DBD1BEDB24EA866595EB5C694EB75","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=593DBD1BEDB24EA866595EB5C694EB75","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80p0z000000n9mq3662A_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80p0z000000n9mq3662A_C_180_180.jpg"},"identityType":13,"tag":""},"images":[{"imageId":309725548,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0h12000000skgniE51D_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0h12000000skgniE51D.jpg","width":600,"height":337,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":208,"likeCount":55,"commentCount":33,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4765591.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4765591.html","wxUrl":"/pages/article/detail?articleId=4765591"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2NTU5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4765591","wxUrl":"/pages/detail/detail?articleId=4765591"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2NTU5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4765591","wxUrl":"/pages/you/lvpai/detail?articleId=4765591"}],"tags":[],"topics":[{"topicId":6544,"topicName":"踏青赏花正当时","level":1},{"topicId":3100,"topicName":"春天去哪玩","level":1}],"pois":[{"poiType":3,"poiId":10547759,"poiName":"婺源篁岭","businessId":141512,"districtId":446,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=10547759&poiType=3&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=10547759&poiType=3&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"},{"poiType":4,"poiId":446,"poiName":"婺源","districtId":446,"districtName":"婺源","districtENName":"Wuyuan","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=446&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=446&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":true,"countryName":"中国"}],"publishTime":"2019-03-20 21:49:14","publishTimeDisplay":"22小时前","shootTime":"2019-03-20 21:39:34","shootTimeDisplay":"2019-03-20","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":0,"articleStatus":2,"poiName":"婺源篁岭"}},{"type":1,"article":{"articleId":4765518,"articleType":"hot","productType":5,"sourceType":126,"articleTitle":"普吉镇最好吃的高颜值椰子冰淇淋&玫瑰水甜品店Torry’s 位于普吉镇彩虹巷里面，一座粉红的房子","author":{"authorId":6829,"nickName":"靓雅雅vivian","clientAuth":"7C03A236C9FA5CDDAA08A15EEE75D36A","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=7C03A236C9FA5CDDAA08A15EEE75D36A","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/fd/headphoto/g2/M08/45/2B/CghzgVWAYNSALZi3AAEcuwmVcpA112_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/fd/headphoto/g2/M08/45/2B/CghzgVWAYNSALZi3AAEcuwmVcpA112_C_180_180.jpg"},"identityType":13,"tag":""},"images":[{"imageId":309724728,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0l12000000sj18i52A7_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0l12000000sj18i52A7.jpg","width":1080,"height":1616,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":78,"likeCount":13,"commentCount":13,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4765518.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4765518.html","wxUrl":"/pages/article/detail?articleId=4765518"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2NTUxOA==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4765518","wxUrl":"/pages/detail/detail?articleId=4765518"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2NTUxOA==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4765518","wxUrl":"/pages/you/lvpai/detail?articleId=4765518"}],"tags":[],"topics":[{"topicId":30,"topicName":"网红打卡地","level":1}],"pois":[{"poiType":2,"poiId":31366362,"poiName":"Torry's Icecream","businessId":15804486,"districtId":364,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=31366362&poiType=2&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=31366362&poiType=2&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":false,"countryName":"泰国"},{"poiType":4,"poiId":364,"poiName":"普吉岛","districtId":364,"districtName":"普吉岛","districtENName":"Phuket","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=364&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=364&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":false,"countryName":"泰国"}],"publishTime":"2019-03-20 21:46:49","publishTimeDisplay":"22小时前","shootTime":"2019-03-16 20:26:48","shootTimeDisplay":"2019-03-16","level":3,"distanceText":"","isLike":false,"imageCounts":8,"isCollected":false,"collectCount":0,"articleStatus":2,"poiName":"Torry's Icecream"}},{"type":1,"article":{"articleId":4374931,"articleType":"individual","productType":5,"sourceType":126,"articleTitle":"人生可以更美！最孤独的图书馆！","author":{"authorId":419232,"nickName":"Chris蕊","clientAuth":"FADD41B0B73CBCED22FD62A7368AC4CBD1A061CCF40F2F329419EEB88B8C2702","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=FADD41B0B73CBCED22FD62A7368AC4CBD1A061CCF40F2F329419EEB88B8C2702","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z8080x000000l1kfd80D5_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z8080x000000l1kfd80D5_C_180_180.jpg"},"tag":""},"images":[{"imageId":304999841,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100d10000000pfn44F55A_R_800_10000_Q50.png","originalUrl":"https://youimg1.c-ctrip.com/target/100d10000000pfn44F55A.png","width":640,"height":960,"mediaType":0}],"hasVideo":false,"readCount":27478,"likeCount":3084,"commentCount":1172,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4374931.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4374931.html","wxUrl":"/pages/article/detail?articleId=4374931"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDM3NDkzMQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4374931","wxUrl":"/pages/detail/detail?articleId=4374931"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDM3NDkzMQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4374931","wxUrl":"/pages/you/lvpai/detail?articleId=4374931"}],"tags":[],"topics":[{"topicId":30,"topicName":"网红打卡地","level":1},{"topicId":1005,"topicName":"旅拍美少女","level":1},{"topicId":429,"topicName":"春节出游","level":3}],"pois":[{"poiType":3,"poiId":30406479,"poiName":"阿那亚","businessId":1984893,"districtId":1835,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=30406479&poiType=3&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=30406479&poiType=3&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"},{"poiType":4,"poiId":1835,"poiName":"昌黎","districtId":1835,"districtName":"昌黎","districtENName":"Changli","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=1835&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=1835&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":true,"countryName":"中国"}],"publishTime":"2019-01-18 16:36:36","publishTimeDisplay":"01-18","shootTime":"2019-01-18 16:36:36","shootTimeDisplay":"2019-01-18","level":2,"distanceText":"","isLike":false,"imageCounts":7,"isCollected":false,"collectCount":88,"articleStatus":2,"poiName":"阿那亚"}},{"type":1,"article":{"articleId":4674889,"articleType":"individual","productType":5,"sourceType":126,"articleTitle":"巴厘岛怎么选择合适的酒店？ 大家可以根据你入住的酒店选择附近的区域游玩，缩短车路时间，让旅行更加明智","author":{"authorId":5049,"nickName":"缅茨沐","clientAuth":"AAA514E0B6B1A6A50E7D5FB1641606E0","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=AAA514E0B6B1A6A50E7D5FB1641606E0","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80g12000000s4f1y3154_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80g12000000s4f1y3154_C_180_180.jpg"},"identityType":13,"tag":""},"images":[{"imageId":308829004,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0e12000000rw2fqD2D5_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0e12000000rw2fqD2D5.jpg","width":1920,"height":1079,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":8853,"likeCount":3830,"commentCount":1334,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4674889.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4674889.html","wxUrl":"/pages/article/detail?articleId=4674889"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDY3NDg4OQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4674889","wxUrl":"/pages/detail/detail?articleId=4674889"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDY3NDg4OQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4674889","wxUrl":"/pages/you/lvpai/detail?articleId=4674889"}],"tags":[],"topics":[{"topicId":4930,"topicName":"女王出巡记","level":1,"iconImage":{"imageId":308625374,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100b0y000000m6mqa07F4_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100b0y000000m6mqa07F4.png","width":48,"height":48,"mediaType":0}},{"topicId":3104,"topicName":"海边拍照攻略","level":1},{"topicId":14,"topicName":"特色住宿","level":1}],"pois":[{"poiType":4,"poiId":438,"poiName":"巴厘岛","districtId":438,"districtName":"巴厘岛","districtENName":"Bali","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=438&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=438&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":false,"countryName":"印度尼西亚"}],"publishTime":"2019-03-10 15:46:22","publishTimeDisplay":"11天前","shootTime":"2019-03-10 15:45:21","shootTimeDisplay":"2019-03-10","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":100,"articleStatus":2,"poiName":""}},{"type":1,"article":{"articleId":4325091,"articleType":"individual","productType":5,"sourceType":126,"articleTitle":"L͟I͟N͟E͟H͟O͟T͟E͟L͟洛杉矶的网红酒店 其实周围去LA的人并不多,所以能做功课的渠道","author":{"authorId":408543,"nickName":"lulul1l","clientAuth":"5DEDDBA50DAB7A488BB3A02FB1B6E2FD","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=5DEDDBA50DAB7A488BB3A02FB1B6E2FD","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80j0z000000nk6zp3051_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80j0z000000nk6zp3051_C_180_180.jpg"},"tag":""},"images":[{"imageId":303966202,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100i10000000oe7ycDE54_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/100i10000000oe7ycDE54.jpg","width":1920,"height":1435,"mediaType":0,"isWaterMarked":true}],"hasVideo":false,"readCount":13809,"likeCount":2301,"commentCount":449,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4325091.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4325091.html","wxUrl":"/pages/article/detail?articleId=4325091"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDMyNTA5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4325091","wxUrl":"/pages/detail/detail?articleId=4325091"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDMyNTA5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4325091","wxUrl":"/pages/you/lvpai/detail?articleId=4325091"}],"tags":[],"topics":[{"topicId":1449,"topicName":"2019心愿之旅","level":1},{"topicId":30,"topicName":"网红打卡地","level":1},{"topicId":1005,"topicName":"旅拍美少女","level":1},{"topicId":4132,"topicName":"拍照攻略大赛","level":1,"iconImage":{"imageId":306610393,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6.png","width":39,"height":45,"mediaType":0}},{"topicId":3096,"topicName":"一学就会拍照技术流","level":1}],"pois":[{"poiType":50,"poiId":21788633,"poiName":"岚酒店","businessId":736156,"districtId":120518,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=21788633&poiType=50&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=21788633&poiType=50&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":false,"countryName":"美国"},{"poiType":4,"poiId":120518,"poiName":"洛杉矶","districtId":120518,"districtName":"洛杉矶","districtENName":"LosAngeles","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=120518&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=120518&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":false,"countryName":"美国"}],"publishTime":"2019-03-09 07:01:07","publishTimeDisplay":"12天前","shootTime":"2019-01-09 10:39:40","shootTimeDisplay":"2019-01-09","level":2,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":34,"articleStatus":2,"poiName":"岚酒店"}},{"type":2,"poi":{"poiType":3,"poiId":87435,"poiName":"厦门园林博览苑","hotValue":102,"article":{"articleId":4511791,"productType":5,"sourceType":126,"articleTitle":"让你不出厦门就能环游世界的厦门小众拍照圣地之园博园 位置：集美中洲岛 关键词：生态湿地公园、温泉度","author":{"authorId":73783,"nickName":"古月初夏","clientAuth":"C5D27CADBF0163FC62F59B66E188D9AE","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=C5D27CADBF0163FC62F59B66E188D9AE","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/Z80t11000000rbqb5BA3F_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/Z80t11000000rbqb5BA3F_C_180_180.jpg"},"identityType":13,"tag":""},"images":[{"imageId":306465604,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0l11000000qv35dD688_R_640_10000_Q90.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0l11000000qv35dD688.jpg","width":1920,"height":1920,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":3319,"likeCount":267,"commentCount":26,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4511791.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4511791.html","wxUrl":"/pages/article/detail?articleId=4511791"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDUxMTc5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4511791","wxUrl":"/pages/detail/detail?articleId=4511791"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDUxMTc5MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4511791","wxUrl":"/pages/you/lvpai/detail?articleId=4511791"}],"tags":[],"topics":[{"topicId":3173,"topicName":"寻找旅途中的爱情","level":1,"iconImage":{"imageId":306374259,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6.png","width":39,"height":45,"mediaType":0}},{"topicId":2668,"topicName":"新年最酷旅行","level":1,"iconImage":{"imageId":305077609,"dynamicUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6_R_640_10000_Q90.png","originalUrl":"https://youimg1.c-ctrip.com/target/100m0z000000nzaz305A6.png","width":39,"height":45,"mediaType":0}},{"topicId":3097,"topicName":"景点拍照指南","level":1},{"topicId":3098,"topicName":"闺蜜拍照攻略","level":1}],"pois":[{"poiType":3,"poiId":87435,"poiName":"厦门园林博览苑","businessId":63494,"districtId":21,"poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=87435&poiType=3&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=87435&poiType=3&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":true,"countryName":"中国"},{"poiType":4,"poiId":21,"poiName":"厦门","districtId":21,"districtName":"厦门","districtENName":"Xiamen","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=21&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=21&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":1,"isInChina":true,"countryName":"中国"}],"publishTime":"2019-02-13 11:50:15","publishTimeDisplay":"02-13","shootTime":"2019-02-13 11:50:15","shootTimeDisplay":"2019-02-13","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":2,"articleStatus":2,"poiName":"厦门园林博览苑"},"distanceText":">30km","tripShootCount":35,"shortFeature":"参观九大不同风格园林","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=87435&poiType=3&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=87435&poiType=3&isHideHeader=true&isHideNavBar=YES"}},{"type":1,"article":{"articleId":4761451,"articleType":"newArticle","productType":5,"sourceType":126,"articleTitle":"【清迈·Bellavita】 来自台湾的专业香氛理疗师Vita 位于清迈素贴山山脚下有一间Bella","author":{"authorId":411067,"nickName":"温度旅行","clientAuth":"08F2EBEB1C368D9C695B307CE57EF2C9","jumpUrl":"https://m.ctrip.com/webapp/you/user?clientAuth=08F2EBEB1C368D9C695B307CE57EF2C9","coverImage":{"dynamicUrl":"https://images4.c-ctrip.com/target/300l0b0000005o16hBE0F_C_180_180.jpg","originalUrl":"https://images4.c-ctrip.com/target/300l0b0000005o16hBE0F_C_180_180.jpg"},"tag":""},"images":[{"imageId":309652032,"dynamicUrl":"https://youimg1.c-ctrip.com/target/1A0r12000000sfuh12B7A_R_800_10000_Q50.jpg","originalUrl":"https://youimg1.c-ctrip.com/target/1A0r12000000sfuh12B7A.jpg","width":1440,"height":1080,"mediaType":0,"isWaterMarked":false}],"hasVideo":false,"readCount":159,"likeCount":11,"commentCount":4,"urls":[{"version":"article","appUrl":"ctrip://wireless/h5?page=index.html#detail/4761451.html&path=destination/article&navBarStyle=white","h5Url":"https://m.ctrip.com/webapp/you/article/detail/4761451.html","wxUrl":"/pages/article/detail?articleId=4761451"},{"version":"tripShoot","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2MTQ1MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4761451","wxUrl":"/pages/detail/detail?articleId=4761451"},{"version":"tripShoot2","appUrl":"ctrip://wireless/h5?type=5&url=L3JuX2Rlc3RpbmF0aW9uX3ZpZGVvL21haW4uanM/Q1JOTW9kdWxlTmFtZT1kZXN0aW5hdGlvbmxpdmUmQ1JOVHlwZT0xJmluaXRpYWxQYWdlPXRyYXZlbFJlY29yZF9kZXRhaWwmaWQ9NDc2MTQ1MQ==","h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/detail.html?Id=4761451","wxUrl":"/pages/you/lvpai/detail?articleId=4761451"}],"tags":[],"topics":[{"topicId":5590,"topicName":"种草网红民宿","level":1}],"pois":[{"poiType":4,"poiId":209,"poiName":"清迈","districtId":209,"districtName":"清迈","districtENName":"ChiangMai","poiExt":{"h5Url":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=209&poiType=4&isHideHeader=true&isHideNavBar=YES","appUrl":"https://m.ctrip.com/webapp/you/livestream/paipai/pk?poiId=209&poiType=4&isHideHeader=true&isHideNavBar=YES"},"source":2,"isMain":0,"isInChina":false,"countryName":"泰国"}],"publishTime":"2019-03-20 10:42:12","publishTimeDisplay":"1天前","shootTime":"2019-01-24 09:02:50","shootTimeDisplay":"2019-01-24","level":3,"distanceText":"","isLike":false,"imageCounts":9,"isCollected":false,"collectCount":2,"articleStatus":2,"poiName":""}}]
 
+///旅拍页模型
 class TravelItemModel {
-  int totalCount;
-  List<TravelItem> resultList;
+  int? totalCount;
+  List<TravelItem>? resultList;
 
   TravelItemModel({this.totalCount, this.resultList});
 
   TravelItemModel.fromMap(Map<String, dynamic> json) {
     totalCount = json['totalCount'];
     if (json['resultList'] != null) {
-      resultList = new List<TravelItem>();
+      resultList = new List<TravelItem>.empty(growable: true);
       json['resultList'].forEach((v) {
-        resultList.add(new TravelItem.fromJson(v));
+        resultList?.add(new TravelItem.fromJson(v));
       });
     }
   }
@@ -22,91 +23,90 @@ class TravelItemModel {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['totalCount'] = this.totalCount;
     if (this.resultList != null) {
-      data['resultList'] = this.resultList.map((v) => v.toJson()).toList();
+      data['resultList'] = this.resultList?.map((v) => v.toJson()).toList();
     }
     return data;
   }
 }
 
 class TravelItem {
-  int type;
-  Article article;
+  late int type;
+  late Article article;
 
-  TravelItem({this.type, this.article});
+  TravelItem(this.type, this.article);
 
   TravelItem.fromJson(Map<String, dynamic> json) {
     type = json['type'];
-    article =
-    json['article'] != null ? new Article.fromJson(json['article']) : null;
+    article = Article.fromJson(json['article']);
   }
 
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = new Map<String, dynamic>();
     data['type'] = this.type;
     if (this.article != null) {
-      data['article'] = this.article.toJson();
+      data['article'] = this.article?.toJson();
     }
     return data;
   }
 }
 
 class Article {
-  int articleId;
-  String articleType;
-  int productType;
-  int sourceType;
-  String articleTitle;
-  Author author;
-  List<Images> images;
-  bool hasVideo;
-  int readCount;
-  int likeCount;
-  int commentCount;
-  List<Urls> urls;
-  List<Null> tags;
-  List<Topics> topics;
-  List<Pois> pois;
-  String publishTime;
-  String publishTimeDisplay;
-  String shootTime;
-  String shootTimeDisplay;
-  int level;
-  String distanceText;
-  bool isLike;
-  int imageCounts;
-  bool isCollected;
-  int collectCount;
-  int articleStatus;
-  String poiName;
+  late int articleId;
+  String? articleType;
+  late int productType;
+  late int sourceType;
+  late String articleTitle;
+  Author? author;
+  late List<Images> images;
+  late bool hasVideo;
+  late int readCount;
+  late int likeCount;
+  late int commentCount;
+  late List<Urls> urls;
+  late List<Null> tags;
+  late List<Topics> topics;
+  List<Pois>? pois;
+  late String publishTime;
+  late String publishTimeDisplay;
+  late String shootTime;
+  late String shootTimeDisplay;
+  late int level;
+  late String distanceText;
+  late bool isLike;
+  late int imageCounts;
+  late bool isCollected;
+  late int collectCount;
+  late int articleStatus;
+  late String poiName;
 
   Article(
-      {this.articleId,
-        this.articleType,
-        this.productType,
-        this.sourceType,
-        this.articleTitle,
-        this.author,
-        this.images,
-        this.hasVideo,
-        this.readCount,
-        this.likeCount,
-        this.commentCount,
-        this.urls,
-        this.tags,
-        this.topics,
-        this.pois,
-        this.publishTime,
-        this.publishTimeDisplay,
-        this.shootTime,
-        this.shootTimeDisplay,
-        this.level,
-        this.distanceText,
-        this.isLike,
-        this.imageCounts,
-        this.isCollected,
-        this.collectCount,
-        this.articleStatus,
-        this.poiName});
+      this.articleId,
+      this.articleType,
+      this.productType,
+      this.sourceType,
+      this.articleTitle,
+      this.author,
+      this.images,
+      this.hasVideo,
+      this.readCount,
+      this.likeCount,
+      this.commentCount,
+      this.urls,
+      this.tags,
+      this.topics,
+      this.pois,
+      this.publishTime,
+      this.publishTimeDisplay,
+      this.shootTime,
+      this.shootTimeDisplay,
+      this.level,
+      this.distanceText,
+      this.isLike,
+      this.imageCounts,
+      this.isCollected,
+      this.collectCount,
+      this.articleStatus,
+      this.poiName);
 
   Article.fromJson(Map<String, dynamic> json) {
     articleId = json['articleId'];
@@ -115,11 +115,11 @@ class Article {
     sourceType = json['sourceType'];
     articleTitle = json['articleTitle'];
     author =
-    json['author'] != null ? new Author.fromJson(json['author']) : null;
+        json['author'] != null ? new Author.fromJson(json['author']) : null;
     if (json['images'] != null) {
-      images = new List<Images>();
+      images = new List<Images>.empty(growable: true);
       json['images'].forEach((v) {
-        images.add(new Images.fromJson(v));
+        images?.add(new Images.fromJson(v));
       });
     }
     hasVideo = json['hasVideo'];
@@ -127,21 +127,21 @@ class Article {
     likeCount = json['likeCount'];
     commentCount = json['commentCount'];
     if (json['urls'] != null) {
-      urls = new List<Urls>();
+      urls = new List<Urls>.empty(growable: true);
       json['urls'].forEach((v) {
-        urls.add(new Urls.fromJson(v));
+        urls?.add(new Urls.fromJson(v));
       });
     }
     if (json['topics'] != null) {
-      topics = new List<Topics>();
+      topics = new List<Topics>.empty(growable: true);
       json['topics'].forEach((v) {
-        topics.add(new Topics.fromJson(v));
+        topics?.add(new Topics.fromJson(v));
       });
     }
     if (json['pois'] != null) {
-      pois = new List<Pois>();
+      pois = new List<Pois>.empty(growable: true);
       json['pois'].forEach((v) {
-        pois.add(new Pois.fromJson(v));
+        pois?.add(new Pois.fromJson(v));
       });
     }
     publishTime = json['publishTime'];
@@ -166,23 +166,23 @@ class Article {
     data['sourceType'] = this.sourceType;
     data['articleTitle'] = this.articleTitle;
     if (this.author != null) {
-      data['author'] = this.author.toJson();
+      data['author'] = this.author?.toJson();
     }
     if (this.images != null) {
-      data['images'] = this.images.map((v) => v.toJson()).toList();
+      data['images'] = this.images?.map((v) => v.toJson()).toList();
     }
     data['hasVideo'] = this.hasVideo;
     data['readCount'] = this.readCount;
     data['likeCount'] = this.likeCount;
     data['commentCount'] = this.commentCount;
     if (this.urls != null) {
-      data['urls'] = this.urls.map((v) => v.toJson()).toList();
+      data['urls'] = this.urls?.map((v) => v.toJson()).toList();
     }
     if (this.topics != null) {
-      data['topics'] = this.topics.map((v) => v.toJson()).toList();
+      data['topics'] = this.topics?.map((v) => v.toJson()).toList();
     }
     if (this.pois != null) {
-      data['pois'] = this.pois.map((v) => v.toJson()).toList();
+      data['pois'] = this.pois?.map((v) => v.toJson()).toList();
     }
     data['publishTime'] = this.publishTime;
     data['publishTimeDisplay'] = this.publishTimeDisplay;
@@ -201,22 +201,16 @@ class Article {
 }
 
 class Author {
-  int authorId;
-  String nickName;
-  String clientAuth;
-  String jumpUrl;
-  CoverImage coverImage;
-  int identityType;
-  String tag;
+  late int authorId;
+  late String nickName;
+  late String clientAuth;
+  late String jumpUrl;
+  late CoverImage? coverImage;
+  late int identityType;
+  late String tag;
 
-  Author(
-      {this.authorId,
-        this.nickName,
-        this.clientAuth,
-        this.jumpUrl,
-        this.coverImage,
-        this.identityType,
-        this.tag});
+  Author(this.authorId, this.nickName, this.clientAuth, this.jumpUrl,
+      this.coverImage, this.identityType, this.tag);
 
   Author.fromJson(Map<String, dynamic> json) {
     authorId = json['authorId'];
@@ -237,7 +231,7 @@ class Author {
     data['clientAuth'] = this.clientAuth;
     data['jumpUrl'] = this.jumpUrl;
     if (this.coverImage != null) {
-      data['coverImage'] = this.coverImage.toJson();
+      data['coverImage'] = this.coverImage?.toJson();
     }
     data['identityType'] = this.identityType;
     data['tag'] = this.tag;
@@ -246,10 +240,10 @@ class Author {
 }
 
 class CoverImage {
-  String dynamicUrl;
-  String originalUrl;
+  late String dynamicUrl;
+  late String originalUrl;
 
-  CoverImage({this.dynamicUrl, this.originalUrl});
+  CoverImage(this.dynamicUrl, this.originalUrl);
 
   CoverImage.fromJson(Map<String, dynamic> json) {
     dynamicUrl = json['dynamicUrl'];
@@ -265,22 +259,16 @@ class CoverImage {
 }
 
 class Images {
-  int imageId;
-  String dynamicUrl;
-  String originalUrl;
-  double width;
-  double height;
-  int mediaType;
-  bool isWaterMarked;
+  late int imageId;
+  late String dynamicUrl;
+  late String originalUrl;
+  late double width;
+  late double height;
+  late int mediaType;
+  late bool isWaterMarked;
 
-  Images(
-      {this.imageId,
-        this.dynamicUrl,
-        this.originalUrl,
-        this.width,
-        this.height,
-        this.mediaType,
-        this.isWaterMarked});
+  Images(this.imageId, this.dynamicUrl, this.originalUrl, this.width,
+      this.height, this.mediaType, this.isWaterMarked);
 
   Images.fromJson(Map<String, dynamic> json) {
     imageId = json['imageId'];
@@ -306,12 +294,12 @@ class Images {
 }
 
 class Urls {
-  String version;
-  String appUrl;
-  String h5Url;
-  String wxUrl;
+  late String version;
+  late String appUrl;
+  late String h5Url;
+  late String wxUrl;
 
-  Urls({this.version, this.appUrl, this.h5Url, this.wxUrl});
+  Urls(this.version, this.appUrl, this.h5Url, this.wxUrl);
 
   Urls.fromJson(Map<String, dynamic> json) {
     version = json['version'];
@@ -331,11 +319,11 @@ class Urls {
 }
 
 class Topics {
-  int topicId;
-  String topicName;
-  int level;
+  late int topicId;
+  late String topicName;
+  late int level;
 
-  Topics({this.topicId, this.topicName, this.level});
+  Topics(this.topicId, this.topicName, this.level);
 
   Topics.fromJson(Map<String, dynamic> json) {
     topicId = json['topicId'];
@@ -353,28 +341,19 @@ class Topics {
 }
 
 class Pois {
-  int poiType;
-  int poiId;
-  String poiName;
-  int businessId;
-  int districtId;
-  PoiExt poiExt;
-  int source;
-  int isMain;
-  bool isInChina;
-  String countryName;
+  late int poiType;
+  late int poiId;
+  late String poiName;
+  late int businessId;
+  late int districtId;
+  late PoiExt? poiExt;
+  late int source;
+  late int isMain;
+  late bool isInChina;
+  late String countryName;
 
-  Pois(
-      {this.poiType,
-        this.poiId,
-        this.poiName,
-        this.businessId,
-        this.districtId,
-        this.poiExt,
-        this.source,
-        this.isMain,
-        this.isInChina,
-        this.countryName});
+  Pois(this.poiType, this.poiId, this.poiName, this.businessId, this.districtId,
+      this.poiExt, this.source, this.isMain, this.isInChina, this.countryName);
 
   Pois.fromJson(Map<String, dynamic> json) {
     poiType = json['poiType'];
@@ -383,7 +362,7 @@ class Pois {
     businessId = json['businessId'];
     districtId = json['districtId'];
     poiExt =
-    json['poiExt'] != null ? new PoiExt.fromJson(json['poiExt']) : null;
+        json['poiExt'] != null ? new PoiExt.fromJson(json['poiExt']) : null;
     source = json['source'];
     isMain = json['isMain'];
     isInChina = json['isInChina'];
@@ -398,7 +377,7 @@ class Pois {
     data['businessId'] = this.businessId;
     data['districtId'] = this.districtId;
     if (this.poiExt != null) {
-      data['poiExt'] = this.poiExt.toJson();
+      data['poiExt'] = this.poiExt?.toJson();
     }
     data['source'] = this.source;
     data['isMain'] = this.isMain;
@@ -409,10 +388,10 @@ class Pois {
 }
 
 class PoiExt {
-  String h5Url;
-  String appUrl;
+  late String h5Url;
+  late String appUrl;
 
-  PoiExt({this.h5Url, this.appUrl});
+  PoiExt(this.h5Url, this.appUrl);
 
   PoiExt.fromJson(Map<String, dynamic> json) {
     h5Url = json['h5Url'];
